+++ vmm/src/vm.rs
@@ -32,9 +32,9 @@ use crate::{
 };
 use anyhow::anyhow;
 #[cfg(target_arch = "aarch64")]
-use arch::aarch64::gic::gicv3_its::kvm::{KvmGicV3Its, GIC_V3_ITS_SNAPSHOT_ID};
-#[cfg(target_arch = "aarch64")]
 use arch::aarch64::gic::kvm::create_gic;
+#[cfg(target_arch = "aarch64")]
+use arch::aarch64::gic::kvm::{KvmGicV3Its, GIC_V3_ITS_SNAPSHOT_ID};
 use arch::get_host_cpu_phys_bits;
 #[cfg(target_arch = "x86_64")]
 use arch::layout::{KVM_IDENTITY_MAP_START, KVM_TSS_START};
